OpenMP index variables must be signed.

+++ src/openMVG/voctree/vocabulary_tree.hpp
@@ -192,7 +192,7 @@ std::vector<Word> VocabularyTree<Feature, Distance, FeatureAllocator>::quantize(
 
   // quantize the features
   #pragma omp parallel for
-  for(size_t j = 0; j < features.size(); ++j)
+  for(ptrdiff_t j = 0; j < static_cast<ptrdiff_t>(features.size()); ++j)
   {
     // store the visual word associated to the feature in the temporary list
     imgVisualWords[j] = quantize<DescriptorT>(features[j]);
